pyproject.toml
[project]
name = "canslim"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "flask-apscheduler>=1.13.1",
    "flask-cors>=6.0.2",
    "google-adk>=1.25.0",
    "litellm>=1.81.10",
    "openai>=2.20.0",
    "pandas-market-calendars>=5.3.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "supabase>=2.28.0",
    "werkzeug>=3.1.5",
]
